Add tdd for Balanced layout

+++ rosen/modules/2d_graphics/include/text/common_utils.h
@@ -30,12 +30,13 @@ namespace Drawing {
 {
     uint32_t n = 0;
     for (uint32_t i = 0; i < len; i++) {
-        uint32_t c = text[i];
+        uint32_t c = static_cast<uint8_t>(text[i]);
         if (c <= 0x7F) { // 0x00 and 0x7F is the range of utf-8
             n = 0;
         } else if ((c & 0xE0) == 0xC0) { // 0xE0 and 0xC0 is the range of utf-8
             n = 1;
-        } else if (c == 0xED && i < (len - 1) && (text[i + 1] & 0xA0) == 0xA0) { // 0xA0 and 0xED is the range of utf-8
+        } else if (c == 0xED && i < (len - 1) &&
+            (static_cast<uint8_t>(text[i + 1]) & 0xA0) == 0xA0) { // 0xA0 and 0xED is the range of utf-8
             return false;
         } else if ((c & 0xF0) == 0xE0) { // 0xE0 and 0xF0 is the range of utf-8
             n = 2;                       // 2 means the size of range
